First passing test and initial framework
This sets up the initial necessary framework. A lot of this
basically amounts to boilerplate but the main important parts are:

* gdemo/__init__.py contains the WSGI application
* gdemo/urls.map is a file mapping URLs to handlers
* gdemo/tests/test_gabbi.py is the file which generates tests using the
  load_tests protocol supported by unittest-derived test runners
* gdemo/tests/gabbits/front.yaml is one gabbi test file containing one
  working test

test_gabbi.py is the central piece. It establishes where the yaml
files can be found, the wsgi app that is being tested and later will
point at fixtures which can be used with tests.

Note that this infrastructure is not necessary if not needed. You
can choose to run some yaml against a running server with gabbi-run:
http://gabbi.readthedocs.org/en/latest/runner.html