Make the frontpage have some content
A static file is read out of the gdemo package and provided as a
simple front page. A test is added to confirm that the frontpage
cannot be deleted.

Note the difference between the two tests in front.yaml: the first
uses built in defaults for `method` (GET) and `status` (200).

A copy of selector.py has been moved into the local code because the
distributed package is not quite right for Python 3.