Add support of PUT to a container
This provides two functions: update an existing container's owner or
create a container for which we know the ID[1].

Though they have already been used in prior tests, it is worth noting
the $SCHEME and $NETLOC variables. These are replaced with the current
http or https scheme and the host and possible the port of the URL.
These can be used to construct urls in various places. The values are
generated from the WSGI environment which is either provided by a
running server or wsgi-intercept.

[1] The idea being that the container already exists in a notional
space so a PUT is required because we know its URI and a POST should
only be used when we don't.