Implement creating and listing objects
The change makes it possible to POST or PUT an object into a
container and create a list of those objects in a container.

The main new gabbi feature being demonstrated here is reading body
data from a file with <@ signal. In this case it is an image of a
kitten because all demos need kittens.

Note that indexes on JSONPath lists can be - if required.