Implement deleting objects
DELETE on an object will remove it.

The new gabbi feature here is $RESPONSE['$.objects[0]']. This says
"in the body of the prior response get the value at that JSONPath".
In this case that means get me the id of the first item in the
object list. So I can delete it.

There's also a common idiom being used: checking the first and last
items of a list to see if they are the same as a way of saying
"there's only one thing in this list".