Demonstrate fixtures
A fixture in gabbi is per test file. start is called before the
first test and stop is called either after the last test or after
something blows up.

This fixture is is creating a list of objects in a container called
'fixtured' so we can do some later experimenting with that data.

The new test confirms that the right data is created. Note the use
of '$.objects' with a value of a YAML list.