Add a coverage tox target
Having coverage allows us to find where there is missing coverage.
In this case there were several issues with some of the error paths
in object management. There was also an actual bug that fixing the
coverage managed to reveal.

This commit also demonstrates one of the common dangers with the
YAML files used with gabbi. Once you've made it past the elegant and
fine fun of doing test driven development and on into the
maintenance, coverage and regression management phase it is easy to
just thrown more YAML into the same file. Readability is damaged
when this happens.

It's done in this case because on a multi-core machine each yaml
file may run in a different process and the objects.yaml tests all
work from the same data store that is in RAM.

Now that we have reasonable assurance that we have good coverage on
the gdemo/__init.py we can clean it up a bit without too much fear.